// src/app/core/types-enums/leave-request-action.enum.ts
export enum LeaveRequestActionEnum {
  EDIT_DRAFT = "EDIT_DRAFT",
  DELETE_DRAFT = "DELETE_DRAFT",
  SAVE_DRAFT = "SAVE_DRAFT",
  SUBMIT_REQUEST = "SUBMIT_REQUEST",
  CANCEL_REQUEST = "CANCEL_REQUEST",
  APPROVE_REQUEST = "APPROVE_REQUEST",
  REJECT_REQUEST = "REJECT_REQUEST",
  REQUEST_CANCEL = "REQUEST_CANCEL",
  APPROVE_CANCEL = "APPROVE_CANCEL",
  REJECT_CANCEL = "REJECT_CANCEL"
}